requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0
django-cloudinary-storage==0.3.0
idna==3.11
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
